사원 삭제 세부 로직 추가

+++ src/main/java/com/woosan/hr_system/employee/service/EmployeeServiceImpl.java
@@ -108,9 +108,28 @@ public class EmployeeServiceImpl implements EmployeeService {
     }
 
     @Override // 사원 정보 삭제
-    public void deleteEmployee(String employeeId) {
-        employeeDAO.deleteEmployee(employeeId);
-        resignationDAO.deleteResignation(employeeId);
+    public String deleteEmployee(String employeeId) {
+        Employee employee = employeeDAO.getEmployeeById(employeeId);
+        if (employee == null) {
+            return "null"; // 사원 정보 없음
+        } else {
+            employee.setResignation(resignationDAO.getResignedEmployee(employeeId));
+            if (employee.getResignation() == null) {
+                return "null_resignation"; // 퇴사 정보 없음
+            }
+
+            LocalDate resignationDate = employee.getResignation().getResignationDate(); // 퇴사일자
+            LocalDate oneYearLater = resignationDate.plusDays(365); // 퇴사일자 1년 경과일
+            LocalDate now = LocalDate.now();
+
+            if (oneYearLater.isBefore(now) || oneYearLater.isEqual(now)) {
+                employeeDAO.deleteEmployee(employeeId);
+                resignationDAO.deleteResignation(employeeId);
+                return "success"; // 1년이 지남 (삭제 가능)
+            } else {
+                return "not_expired"; // 1년이 지나지 않음
+            }
+        }
     }
 
     @Override // 퇴사 예정인 사원 정보 조회
